# resources/api/openapi.yml
openapi: 3.0.3
info:
  version: 1.0.0-beta
  title: CritterSystem API
  description: >
    The CritterSystem API provides comprehensive access to event management functionality.
    This API is as stable as a **beta** version might be.
    It could burst into flames and morph into a monster at any second!
    (But we try to keep it somewhat consistent, at least during events).
  contact:
    name: GitHub Issues
    url: https://github.com/engelsystem/engelsystem/issues
  license:
    name: GPL 2.0
    # identifier: GPL-2.0
servers:
  - url: /api/v0-beta
    description: This server (Legacy v0-beta)
  - url: /api/v2
    description: This server (v2)
  - url: '{host}/api/v0-beta'
    description: CritterSystem instance to use (Legacy v0-beta)
    variables:
      host:
        default: http://localhost:5080
  - url: '{host}/api/v2'
    description: CritterSystem instance to use (v2)
    variables:
      host:
        default: http://localhost:5080
...
security:
  - bearer-auth: [ ]
  - api-key-header: [ ]
components:
  securitySchemes:
    bearer-auth:
      type: http
      scheme: bearer
      bearerFormat: API key from settings
    api-key-header:
      type: apiKey
      name: x-api-key
      in: header
  responses:
    UnauthorizedError: # 401
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFoundError: # 404
      description: This resource can not be found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Department:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: Security Team
        description:
          type: string
          example: Event security and safety coordination
        contact_info:
          type: object
          properties:
            email:
              type: string
              nullable: true
              example: security@example.com
            phone:
              type: string
              nullable: true
              example: +1234567890
        is_active:
          type: boolean
          example: true
        created_at:
          $ref: '#/components/schemas/DateTime'
        updated_at:
          $ref: '#/components/schemas/DateTimeOptional'
      required:
        - uuid
        - name
        - description
        - is_active
        - created_at
        - updated_at
...
    DateTime:
      type: string
      format: date-time
      description: DateTime in ISO-8601 format
      example: 2023-05-23T00:00:00.000000Z
    DateTimeOptional:
      type: string
      format: date-time
      description: DateTime in ISO-8601 format
      example: 2023-05-23T00:00:00.000000Z
      nullable: true
    Error:
      type: object
      properties:
        message:
          type: string
...
paths:
...
  /departments:
    get:
      tags:
        - department
      summary: Get list of departments
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Department'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /departments/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        description: The department UUID
        schema:
          type: string
          format: uuid
    get:
      tags:
        - department
      summary: Get department details
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Department'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'